[NCL-9124] Rename endpoint /builds/brew-push/{id} to /build-pushes/{id}
