+++ rest-api/pnc-openapi.yaml
@@ -1550,6 +1550,34 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ErrorResponse'
+  /build-pushes/{id}:
+    get:
+      tags:
+      - Build pushes
+      summary: Get specific Report of Build Push Operation.
+      operationId: getPushReport
+      parameters:
+      - name: id
+        in: path
+        description: ID of the Build Push Report.
+        required: true
+        schema:
+          type: string
+      responses:
+        "200":
+          description: Success with results
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/BuildPushReport'
+        "404":
+          description: Can not find specified result
+        "500":
+          description: Server error
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/ErrorResponse'
   /builds:
     get:
       tags:
@@ -1624,34 +1652,6 @@ paths:
             application/json:
               schema:
                 $ref: '#/components/schemas/ErrorResponse'
-        "500":
-          description: Server error
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/ErrorResponse'
-  /builds/brew-push/{id}:
-    get:
-      tags:
-      - Builds
-      summary: Get specific report of build push operation.
-      operationId: getPushReport
-      parameters:
-      - name: id
-        in: path
-        description: ID of the build push operation
-        required: true
-        schema:
-          type: string
-      responses:
-        "200":
-          description: Success with results
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/BuildPushReport'
-        "404":
-          description: Can not find specified result
         "500":
           description: Server error
           content:
